[다이어그램 만들기 전 커밋,푸쉬]
잔여 포인트, 포인트 리스트 출력 mapper 작성만 남음
