--- src/main/java/com/kh/samdado/common/model/Income.java
@@ -1,19 +1,0 @@
-package com.kh.samdado.common.model;
-
-import java.sql.Date;
-
-public class Income {
-//	PAY_NO	NUMBER
-//	ITEM	VARCHAR2(30 BYTE)
-//	AMOUNT	VARCHAR2(30 BYTE)
-//	PDATE	DATE
-//	EXDATE	DATE
-//	US_NO	VARCHAR2(15 BYTE)
-
-	private int payno;
-	private String item;
-	private String amount;
-	private Date pdate;
-	private Date exdate;
-	private String usno;
-}
